LayoutTests/imported/w3c/web-platform-tests/css/css-contain/container-queries/container-longhand-animation-type-expected.txt


PASS Reference variable is applied
PASS container-name is not animatable
PASS container-type is not animatable
